Refactor the SCPI helper methods into suitable helper classes
